+++ vortex-file/src/lib.rs
@@ -98,6 +98,7 @@ pub mod io;
 mod open;
 pub mod read;
 pub mod segments;
+mod strategy;
 #[cfg(test)]
 mod tests;
 mod writer;
