+++ vortex-layout/src/layouts/flat/eval_expr.rs
@@ -80,7 +80,7 @@ mod test {
 
     use crate::layouts::flat::writer::FlatLayoutWriter;
     use crate::segments::test::TestSegments;
-    use crate::strategies::LayoutWriterExt;
+    use crate::writer::LayoutWriterExt;
 
     #[test]
     fn flat_identity() {
